sort nodes

+++ bigsort.py
@@ -6,22 +6,25 @@ import argparse
 import random
 import tempfile
 import math
-from logzero import logger
 
-# logger.info()
-import logging
-logger = logging.getLogger()
-logger.propagate = False
-logger.handlers.clear()
-logger.setLevel(level=logging.INFO)
-# logger.setLevel(level=logging.ERROR)
-# handler = logging.FileHandler("log.txt")
-handler = logging.StreamHandler()
-handler.setLevel(logging.INFO)
-formatter = logging.Formatter(
-    '%(asctime)s %(filename)s-%(lineno)d-%(levelname)s - %(message)s')
-handler.setFormatter(formatter)
-logger.addHandler(handler)
+import logzero
+from logzero import logger
+# logzero.loglevel(logzero.INFO)
+logzero.loglevel(logzero.ERROR)
+# logger.setLevel(level=logger.INFO)
+# import logging
+# logger = logging.getLogger()
+# logger.propagate = False
+# logger.handlers.clear()
+# logger.setLevel(level=logging.INFO)
+# # logger.setLevel(level=logging.ERROR)
+# # handler = logging.FileHandler("log.txt")
+# handler = logging.StreamHandler()
+# handler.setLevel(logging.INFO)
+# formatter = logging.Formatter(
+#     '%(asctime)s %(filename)s-%(lineno)d-%(levelname)s - %(message)s')
+# handler.setFormatter(formatter)
+# logger.addHandler(handler)
 
 
 def free():
@@ -88,18 +91,24 @@ class Node:
         self.n_line = len(batch)
         self.head = batch[0]
         self.tail = batch[-1]
-        writer = open(name, mode='w', buffering=buffering)
-        writer.writelines(batch)
-        writer.close()
+        if name==None:
+            self.bucket=batch
+        else:
+            writer = open(name, mode='w', buffering=buffering)
+            writer.writelines(batch)
+            writer.close()
 
     def catch(self):
+        if self.name==None:
+            return self.bucket
         bucket = open(self.name).readlines()
         os.unlink(self.name)
         return bucket
 
 
 class BigSort:
-    def __init__(self, sortType='i', unique=False, budget=0.8, nSplit=10, nLine=100000, buffering=1024*1024, tmpDir=None, sortFn=sortFn, splitFn=splitFn):
+    def __init__(self, sortType='i', unique=False,head=-1, budget=0.8, nSplit=10, nLine=100000, buffering=1024*1024, tmpDir=None, sortFn=sortFn, splitFn=splitFn):
+        self.head = head
         self.buffering = buffering
         self.nSplit = nSplit
         self.nLine = nLine
@@ -114,10 +123,11 @@ class BigSort:
         self.splitFn = splitFn
         self.MEM = free()*budget
         logger.info(f"MEM  {self.MEM//1024//1024}M")
+        self.n_readed=0
+        self.n_writed=0
 
     def map(self, reader, folder):
         Nodes = []
-        block = None
         bucket = []
         total = 0
         n_block = 0
@@ -128,58 +138,52 @@ class BigSort:
                 if usage < self.MEM:
                     continue
                 total += len(bucket)
-                if not block:
-                    block_dir = f"{folder}/block-{n_block}"
-                    block = Block(block_dir, bucket, self.sortType, self.nSplit, self.buffering, self.sortFn)
-                    logger.info(f"{total} {len(bucket)} --> {block_dir}")
-                    Nodes += block.Nodes
+                block_dir = f"{folder}/block-{n_block}"
+                block = Block(block_dir, bucket, self.sortType, self.nSplit, self.buffering, self.sortFn)
+                logger.info(f"map:{total} bucket:{len(bucket)} --> {block_dir}")
+                Nodes += block.Nodes
                 bucket = []
                 n_block += 1
-                block = None
 
         if bucket:
             bucket = sortFn(bucket, self.sortType)
             if total == 0:
-                return Nodes, bucket
-            total += len(bucket)
-            if not block:
+                node=Node(None,bucket,self.buffering)
+                Nodes.append(node)
+                total += len(bucket)
+            else:
+                total += len(bucket)
                 block_dir = f"{folder}/block-{n_block}"
                 block = Block(block_dir, bucket, self.sortType, self.nSplit, self.buffering, self.sortFn)
-                logger.info(f"{total} {len(bucket)} --> {block_dir}")
+                logger.info(f"map:{total} bucket:{len(bucket)} --> {block_dir}")
                 Nodes += block.Nodes
-        logger.info(f"map done!  {vars(reader)}  {total}  -->  {len(Nodes)} ")
-        return Nodes, bucket
+        logger.info(f"map done!  {vars(reader)}  maped:{total}  -->  Nodes:{len(Nodes)} ")
+        return Nodes
 
     def write(self, doc, writer, last=None):
-        if not self.unique:
-            writer.writelines(doc)
-            return
+        self.n_readed+=len(doc)
         for x in doc:
-            if x == last:
+            if self.n_writed>=self.head>=0:
+                return 
+            if self.unique and  x == last:
                 continue
             writer.write(x)
             last = x
+            self.n_writed+=1
         return last
 
-    def reduce(self, Nodes, writer, bucket):
+    def reduce(self, Nodes, writer):
         queue = []
         n_read = 0
         n_write = 0
         last = None
 
-        if not Nodes:
-            n_read=len(bucket)
-            last = self.write(bucket, writer, last)
-            logger.info(f"reduce done! {n_read} --> {n_write} {vars(writer)}")
-            return
-
-        Nodes = [(x.head, x) for x in Nodes]
         Nodes = self.sortFn(Nodes, self.sortType)
         for i, x in enumerate(Nodes):
-            head, node = x
+            node = x[-1]
             bucket = node.catch()
             queue += bucket
-            pivot = Nodes[i+1][1].head if i != len(Nodes)-1 else None
+            pivot = Nodes[i+1][-1].head if i != len(Nodes)-1 else None
             queue = self.sortFn(queue, self.sortType)
             lines, queue = self.splitFn(queue, self.sortType, pivot, self.nSplit)
             last = self.write(lines, writer, last)
@@ -188,7 +192,7 @@ class BigSort:
             n_read += r1
             n_write += w1
             if i % self.nSplit == 0:
-                logger.info(f"{i}/{len(Nodes)} {n_read} --{n_write} queue:{len(queue)}  {r1} -> {w1}   {r1-w1} ")
+                logger.info(f"Node:{i}/{len(Nodes)} n_read:{n_read} -- n_write:{n_write} queue:{len(queue)}  r1:{r1} -> w1:{w1}   {r1-w1} ")
 
         if queue:
             pivot = None
@@ -199,29 +203,31 @@ class BigSort:
             w1 = len(lines)
             n_read += r1
             n_write += w1
-            logger.info(f"{i}/{len(Nodes)} {n_read} --{n_write}  {r1} -> {w1}   {r1-w1} ")
+            logger.info(f"Node:{i}/{len(Nodes)} n_read:{n_read} -- n_write:{n_write} queue:{len(queue)}  r1:{r1} -> w1:{w1}   {r1-w1} ")
 
-        logger.info(f"reduce done! {n_read} --> {n_write} {vars(writer)}")
+        logger.info(f"reduce done! n_read:{n_read} --> n_write:{n_write} {vars(writer)}")
 
     def sort(self, reader, writer):
         # import time
         # t0=time.time()
         temp_dir = tempfile.TemporaryDirectory(dir=self.tmpDir)
-        Nodes, bucket = self.map(reader, temp_dir.name)
+        Nodes = self.map(reader, temp_dir.name)
+        Nodes = [(x.head, x.tail,i,x) for i,x in enumerate(Nodes)]
         # t1=time.time()
-        self.reduce(Nodes, writer, bucket)
-        # temp_dir.cleanup()
+        self.reduce(Nodes, writer)
+        temp_dir.cleanup()
         # t2=time.time()
         # logger.info(f"map {t1-t0} reduce {t2-t1}")
+        logger.info(f" n_readed:{self.n_readed} n_writed:{self.n_writed}")
 
 
-def bigsort(reader, writer, sortType='i', unique=False, budget=0.8, nSplit=10, nLine=10000, tmpDir=None, sortFn=sortFn, splitFn=splitFn):
-    sorter = BigSort(sortType=sortType, unique=unique, budget=budget, nSplit=nSplit, nLine=nLine, tmpDir=tmpDir, sortFn=sortFn, splitFn=splitFn)
+def bigsort(reader, writer, sortType='i', unique=False,head=-1, budget=0.8, nSplit=10, nLine=10000, tmpDir=None, sortFn=sortFn, splitFn=splitFn):
+    sorter = BigSort(sortType=sortType, unique=unique,head=head, budget=budget, nSplit=nSplit, nLine=nLine, tmpDir=tmpDir, sortFn=sortFn, splitFn=splitFn)
     sorter.sort(reader, writer)
     writer.flush()
 
 
-def sortFile(src=None, tgt=None, sortType='i', unique=False, budget=0.8, nSplit=10, nLine=10000, tmpDir=None, buffering=1024*1024):
+def sortFile(src=None, tgt=None, sortType='i', unique=False,head=-1, budget=0.8, nSplit=10, nLine=10000, tmpDir=None, buffering=1024*1024):
     if not src:
         reader = sys.stdin
     else:
@@ -232,8 +238,8 @@ def sortFile(src=None, tgt=None, sortType='i', unique=False, budget=0.8, nSplit=
         writer = sys.stdout
     else:
         writer = open(tgt, 'w', buffering=buffering)
-    bigsort(reader, writer, sortType=sortType, unique=unique, budget=budget, nSplit=nSplit, nLine=nLine, tmpDir=tmpDir)
-    writer.close()
+    bigsort(reader, writer, sortType=sortType, unique=unique,head=head, budget=budget, nSplit=nSplit, nLine=nLine, tmpDir=tmpDir)
+    # writer.close()
 
 
 # https://docs.python.org/3/library/operator.html
@@ -271,9 +277,10 @@ def main():
     parser.add_argument("--buffering", type=int, default=1024*1024)
     parser.add_argument("--nSplit", type=int, default=10)  # bigger if skew or shuffle
     parser.add_argument("--nLine", type=int, default=100000)
-    parser.add_argument("--budget", type=float, default=0.5)  # 0.8 memary budget in ratio
+    parser.add_argument("-m","--budget", type=float, default=0.5)  # 0.8 memary budget in ratio
     parser.add_argument("-u", "--unique", default=False)  # only when sort
     parser.add_argument("-s", "--sortType", default="i")  # one of  'i/d/R': increase descend random
+    parser.add_argument("-n","--number", type=int,default=-1)  # number from head
     parser.add_argument("-T", "--tmpDir", default=None)  # None "_tmp_"
     parser.add_argument("-c", "--checkOrdering")  # check file order
     args = parser.parse_args()
@@ -288,7 +295,7 @@ def main():
             reader = os.popen(src)
         check(reader, args.checkOrdering)
     else:
-        sortFile(args.input, args.output, sortType=args.sortType, unique=args.unique, budget=args.budget, tmpDir=args.tmpDir, nSplit=args.nSplit, nLine=args.nLine, buffering=args.buffering)
+        sortFile(args.input, args.output, sortType=args.sortType, unique=args.unique,head=args.number, budget=args.budget, tmpDir=args.tmpDir, nSplit=args.nSplit, nLine=args.nLine, buffering=args.buffering)
     # sortFile("cat bookcorpus.txt","sorted.txt")
     # check(open("sorted.txt"),"<=")
